No issue: Break down sagas into separated files (#4047)
