+++ packages/web-frontend/src/sagas/handlers/resetToHome.js
@@ -1,0 +1,6 @@
+import { put } from 'redux-saga/effects';
+import { goHome } from '../../actions';
+
+export function* resetToHome() {
+  yield put(goHome());
+}
