+++ packages/web-frontend/src/sagas/watchDashboard/index.js
@@ -1,0 +1,1 @@
+export * from './watchViewFetchRequests';
