+++ packages/web-frontend/src/sagas/watchOrgUnit/watchHandleLocationChange.js
@@ -1,0 +1,7 @@
+import { takeLatest } from 'redux-saga/effects';
+import { LOCATION_CHANGE } from '../../actions';
+import { handleLocationChange } from '../handlers';
+
+export function* watchHandleLocationChange() {
+  yield takeLatest(LOCATION_CHANGE, handleLocationChange);
+}
