+++ packages/web-frontend/src/sagas/watchProject/watchRequestProjectAccess.js
@@ -1,0 +1,7 @@
+import { takeLatest } from 'redux-saga/effects';
+import { REQUEST_PROJECT_ACCESS } from '../../actions';
+import { fetchCountryAccessDataIfRequired } from '../handlers';
+
+export function* watchRequestProjectAccess() {
+  yield takeLatest(REQUEST_PROJECT_ACCESS, fetchCountryAccessDataIfRequired);
+}
