+++ packages/web-frontend/src/sagas/watchSearchbar/index.js
@@ -1,0 +1,2 @@
+export * from './watchFetchMoreSearchResults';
+export * from './watchSearchChange';
