+++ packages/web-frontend/src/sagas/watchUser/refreshBrowserWhenFinishingUserSession.js
@@ -1,0 +1,8 @@
+import { takeLatest } from 'redux-saga/effects';
+import { FINISH_USER_SESSION } from '../../actions';
+
+export function* refreshBrowserWhenFinishingUserSession() {
+  yield takeLatest(FINISH_USER_SESSION, () => {
+    window.location.reload();
+  });
+}
