+++ packages/web-frontend/src/sagas/watchUser/watchFindUserCurrentLoggedIn.js
@@ -1,0 +1,28 @@
+import { call, put, takeLatest } from 'redux-saga/effects';
+import { fetchUserLoginSuccess, findUserLoginFailed, FIND_USER_LOGGEDIN } from '../../actions';
+import { request } from '../../utils';
+
+/**
+ * findUserLoggedIn
+ *
+ * Find if any user was logged in and call action on success/fail.
+ *
+ */
+function* findUserLoggedIn(action) {
+  const requestResourceUrl = 'getUser';
+
+  try {
+    const userData = yield call(request, requestResourceUrl);
+    if (userData.name !== 'public') {
+      yield put(fetchUserLoginSuccess(userData.name, userData.email, action.loginType));
+    } else {
+      yield put(findUserLoginFailed());
+    }
+  } catch (error) {
+    yield put(error.errorFunction(error));
+  }
+}
+
+export function* watchFindUserCurrentLoggedIn() {
+  yield takeLatest(FIND_USER_LOGGEDIN, findUserLoggedIn);
+}
