+++ packages/web-frontend/src/sagas/watchUser/watchLogoutSuccess.js
@@ -1,0 +1,7 @@
+import { takeLatest } from 'redux-saga/effects';
+import { FETCH_LOGOUT_SUCCESS } from '../../actions';
+import { resetToHome } from '../handlers';
+
+export function* watchLogoutSuccess() {
+  yield takeLatest(FETCH_LOGOUT_SUCCESS, resetToHome);
+}
